added dynamodb service shell, added API context, updated index renderer to send initialization flags
