+++ internal/routes/static.go
@@ -4,8 +4,10 @@ import (
 	"net/http"
 
 	"github.com/cmarfia/stories-by-iot/internal/template"
+	"github.com/cmarfia/stories-by-iot/internal/server"
 
 	"github.com/labstack/echo"
+	"github.com/pkg/errors"
 )
 
 func registerStaticRoutes(e *echo.Echo) {
@@ -16,5 +18,22 @@ func registerStaticRoutes(e *echo.Echo) {
 }
 
 func renderIndex(c echo.Context) error {
-	return c.Render(http.StatusOK, "index", template.Index{Title: "Story By Iot", Version: 2})
+	apiContext, ok := c.(*server.APIContext)
+	if !ok {
+		return errors.New("routes: could not get API context for route /")
+	}
+
+	l, err := apiContext.DynamoService.GetLibrary()
+	if err != nil {
+		return errors.Wrap(err, "routes: error fetching library")
+	}
+
+	data := template.Index{
+		Title: "Story By Iot",
+		Version: 2,
+		Flags: map[string]interface{}{
+			"library": l,
+		},
+	}
+	return c.Render(http.StatusOK, "index", data)
 }
