+++ internal/server/apiContext.go
@@ -1,0 +1,13 @@
+package server 
+
+import(
+	"github.com/cmarfia/stories-by-iot/internal/dynamodb"
+
+	"github.com/labstack/echo"
+)
+
+// APIContext represents the custom server context for the API
+type APIContext struct {
+	echo.Context
+	DynamoService *dynamodb.Service
+}
