+++ internal/story/libary.go
@@ -1,0 +1,6 @@
+package story
+
+// Library represents a collection of stories without their contents
+type Library struct {
+	Stories []Info `json:"stories"`
+}
